Separate all validationError functions
